+++ src/App/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
  
 import Navigation from '../components/Navigation';
-import LandingPage from '../components/Landing';
+import LandingPage from '../pages/LandingPage';
 import SignUpPage from '../components/SignUp';
 import SignInPage from '../components/SignIn';
 import PasswordForgetPage from '../components/PasswordForget';
